requirements.txt
Flask
faiss-cpu
sentence-transformers
transformers
